Push the geometries in right order and use shape data when needed

+++ src/lib/VSDXContentCollector.h
@@ -57,11 +57,11 @@ public:
   void collectMoveTo(unsigned id, unsigned level, double x, double y);
   void collectLineTo(unsigned id, unsigned level, double x, double y);
   void collectArcTo(unsigned id, unsigned level, double x2, double y2, double bow);
-  void collectNURBSTo(unsigned id, unsigned level, double x2, double y2, unsigned xType, unsigned yType, double degree, std::vector<std::pair<double, double> > controlPoints, std::vector<double> knotVector, std::vector<double> weights);
+  void collectNURBSTo(unsigned id, unsigned level, double x2, double y2, unsigned xType, unsigned yType, unsigned degree, std::vector<std::pair<double, double> > controlPoints, std::vector<double> knotVector, std::vector<double> weights);
   void collectNURBSTo(unsigned id, unsigned level, double x2, double y2, double knot, double knotPrev, double weight, double weightPrev, unsigned dataID);
   void collectPolylineTo(unsigned id , unsigned level, double x, double y, unsigned xType, unsigned yType, std::vector<std::pair<double, double> > &points);
   void collectPolylineTo(unsigned id , unsigned level, double x, double y, unsigned dataID);
-  void collectShapeData(unsigned id, unsigned level, unsigned xType, unsigned yType, double degree, double lastKnot, std::vector<std::pair<double, double> > controlPoints, std::vector<double> knotVector, std::vector<double> weights);
+  void collectShapeData(unsigned id, unsigned level, unsigned xType, unsigned yType, unsigned degree, double lastKnot, std::vector<std::pair<double, double> > controlPoints, std::vector<double> knotVector, std::vector<double> weights);
   void collectShapeData(unsigned id, unsigned level, unsigned xType, unsigned yType, std::vector<std::pair<double, double> > points);
   void collectXFormData(unsigned id, unsigned level, const XForm &xform);
   void collectShapeId(unsigned id, unsigned level, unsigned shapeId);
@@ -86,7 +86,7 @@ private:
   void transformPoint(double &x, double &y);
   void transformAngle(double &angle);
 
-  double _NURBSBasis(unsigned knot, double degree, double point, unsigned controlCount, const std::vector<double> &knotVector);
+  double _NURBSBasis(unsigned knot, unsigned degree, double point, unsigned controlCount, const std::vector<double> &knotVector);
 
   void _flushCurrentPath();
   void _flushCurrentForeignData();
